Add files via upload
Challenge: A file contains the entire English dictionary. Place the file in the same folder as your Python script.  You will then need to create a script that counts how many words are in the dictionary that contain any consecutive 3-letter sequence in your firstname+lastname (for me, any word containing one of ”jas”, ”aso”, ”son”, “onl”, ”nle”, “lew”, “ewi”, and “wis”; the consecutive 3-letter sequences of “jasonlewis”) - the result should be an integer.

You are to ask the user for their full name (no spaces) and save that in the variable (those lines don't count towards the 3 line count).